+++ lib/ui/like/like.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:seed_app/ui/top/user_profile/user_profile.dart';
+import 'package:seed_app/ui/user_top/user_profile.dart';
 
 class LikePageWidget extends ConsumerWidget {
   final scaffoldKey = GlobalKey<ScaffoldState>();
