--- lib/ui/mypage/menu_importer.dart
@@ -1,8 +1,0 @@
-export 'package:seed_app/ui/mypage/menu/footprint.dart';
-export 'package:seed_app/ui/mypage/menu/help.dart';
-export 'package:seed_app/ui/mypage/menu/license.dart';
-export 'package:seed_app/ui/mypage/menu/notification/notification.dart';
-export 'package:seed_app/ui/mypage/menu/setting.dart';
-export 'package:seed_app/ui/mypage/menu/my_fav.dart';
-export 'package:seed_app/ui/mypage/menu/my_like.dart';
-export 'package:seed_app/ui/mypage/my_profile/my_profile.dart';
