+++ lib/ui/mypage/mypage_menu.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:seed_app/test_page.dart';
-import 'package:seed_app/ui/mypage/menu/footprint.dart';
-import 'package:seed_app/ui/mypage/menu/help.dart';
-import 'package:seed_app/ui/mypage/menu/license.dart';
-import 'package:seed_app/ui/mypage/menu/setting.dart';
-import 'package:seed_app/ui/mypage/menu/my_fav.dart';
-import 'package:seed_app/ui/mypage/menu/my_like.dart';
+import 'package:seed_app/ui/user_menu_to_pages/other_pages/footprint.dart';
+import 'package:seed_app/ui/user_menu_to_pages/other_pages/help.dart';
+import 'package:seed_app/ui/user_menu_to_pages/other_pages/license.dart';
+import 'package:seed_app/ui/user_menu_to_pages/other_pages/setting.dart';
+import 'package:seed_app/ui/user_menu_to_pages/other_pages/yourfav.dart';
+import 'package:seed_app/ui/user_menu_to_pages/other_pages/yourlike.dart';
 
-import 'menu/notification/notification.dart';
+import '../user_menu_to_pages/notification/notification.dart';
 
 class MyPageMenu extends StatelessWidget {
   const MyPageMenu({Key? key}) : super(key: key);
