+++ lib/ui/user_menu_to_pages/my_profile/my_profile.dart
@@ -5,11 +5,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:seed_app/locator.dart';
 import 'package:seed_app/models/profile_item_models.dart';
 import 'package:seed_app/models/user_models.dart';
-import 'package:seed_app/ui/mypage/my_profile/mp_introduction.dart';
-import 'package:seed_app/ui/mypage/my_profile/mp_introduction_card.dart';
-import 'package:seed_app/ui/mypage/my_profile/mp_score.dart';
 
-import 'package:seed_app/ui/mypage/my_profile/mp_items_list.dart';
+import 'package:seed_app/ui/user_menu_to_pages/my_profile/mp_introduction.dart';
+import 'package:seed_app/ui/user_menu_to_pages/my_profile/mp_introduction_card.dart';
+
+import 'package:seed_app/ui/user_menu_to_pages/my_profile/mp_items_list.dart';
+import 'package:seed_app/ui/user_menu_to_pages/my_profile/mp_score.dart';
 import 'package:seed_app/controller/user_controller.dart';
 
 /*
@@ -62,32 +63,32 @@ class MyProfilePageWidget extends ConsumerWidget {
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               // プロフ画像
-              MyProfilePictures(
+              ProfilePictures(
                 avatarUrl: _currentUser?.avatarUrl,
               ),
               // 概要欄
-              const MyIntroductionCard(),
+              const IntroductionCard(),
               // プロフィールスコア
-              const MyProfileScore(),
+              const ProfileScore(),
               // 自由記述欄
-              const MyIntroduction(),
+              const Introduction(),
               // 基本情報
-              MyProfileItemsList(
+              ProfileItemsList(
                 itemName: '基本情報',
                 itemsList: profileItem.basicInfo,
               ),
               // 学歴・職種・外見
-              MyProfileItemsList(
+              ProfileItemsList(
                 itemName: '学歴・職種・外見',
                 itemsList: profileItem.socialInfo,
               ),
               // 性格・趣味・生活
-              MyProfileItemsList(
+              ProfileItemsList(
                 itemName: '性格・趣味・生活',
                 itemsList: profileItem.lifeStyleInfo,
               ),
               // 恋愛・結婚について
-              MyProfileItemsList(
+              ProfileItemsList(
                 itemName: '恋愛・結婚について',
                 itemsList: profileItem.viewOfLove,
               )
@@ -101,9 +102,9 @@ class MyProfilePageWidget extends ConsumerWidget {
 
 // 完成---------------------------------------------------------------------------------------------
 // プロフィール画像
-class MyProfilePictures extends ConsumerWidget {
+class ProfilePictures extends ConsumerWidget {
   final String? avatarUrl;
-  const MyProfilePictures({
+  ProfilePictures({
     this.avatarUrl,
     Key? key,
   }) : super(key: key);
@@ -118,7 +119,7 @@ class MyProfilePictures extends ConsumerWidget {
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(10),
           image: DecorationImage(
-              image: avatarUrl == null || avatarUrl == ''
+              image: avatarUrl == null
                   ? Image.asset('assets/images/user1.jpg').image
                   : Image.file(File(avatarUrl!)).image,
               fit: BoxFit.fitWidth),
