+++ lib/ui/user_menu_to_pages/user_menu_importer.dart
@@ -1,0 +1,8 @@
+export 'package:seed_app/ui/user_menu_to_pages/other_pages/footprint.dart';
+export 'package:seed_app/ui/user_menu_to_pages/other_pages/help.dart';
+export 'package:seed_app/ui/user_menu_to_pages/other_pages/license.dart';
+export 'package:seed_app/ui/user_menu_to_pages/notification/notification.dart';
+export 'package:seed_app/ui/user_menu_to_pages/other_pages/setting.dart';
+export 'package:seed_app/ui/user_menu_to_pages/other_pages/yourfav.dart';
+export 'package:seed_app/ui/user_menu_to_pages/other_pages/yourlike.dart';
+export 'package:seed_app/ui/user_menu_to_pages/my_profile/my_profile.dart';
