+++ lib/ui/what's_ui.txt
@@ -1,0 +1,5 @@
+ページごとにディレクトリを分けて作成
+例えば、login_screenというディレクトリを用意し、LoginScreenというWidgetを作成
+
+
+
